requirements.txt
fastapi[standard]
python-dotenv
llama-cpp-python
langchain-community
langchain-core
torch
torchvision
diffusers
transformers
accelerate
Pillow
elevenlabs